CollegeFeedbackAI_Project3/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.25
textblob==0.15.3
